refactor: move asserters to core-test-kit
