+++ packages/language-service/test/test-kit/postcss-node-asserts.ts
@@ -1,33 +1,9 @@
 import type { Invalid } from '@stylable/language-service/dist/lib-new/invalid-node';
-import type * as postcss from 'postcss';
+export { assertAtRule, assertComment, assertDecl, assertRule } from '@stylable/core-test-kit';
 
-export function assertRule(node: any, msg?: string): postcss.Rule {
-    if (node?.type !== 'rule') {
-        throw new Error('expected postcss rule' + (msg ? ` (${msg})` : ''));
-    }
-    return node;
-}
-export function assertAtRule(node: any, msg?: string): postcss.Rule {
-    if (node?.type !== 'atrule') {
-        throw new Error('expected postcss at-rule' + (msg ? ` (${msg})` : ''));
-    }
-    return node;
-}
-export function assertDecl(node: any, msg?: string): postcss.Declaration {
-    if (node?.type !== 'decl') {
-        throw new Error('expected postcss declaration' + (msg ? ` (${msg})` : ''));
-    }
-    return node;
-}
 export function assertInvalid(node: any, msg?: string): Invalid {
     if (node?.type !== 'invalid') {
         throw new Error('expected invalid node' + (msg ? ` (${msg})` : ''));
-    }
-    return node;
-}
-export function assertComment(node: any, msg?: string): postcss.Comment {
-    if (node?.type !== 'comment') {
-        throw new Error('expected comment node' + (msg ? ` (${msg})` : ''));
     }
     return node;
 }
